Initial commit of sparta.Discovery()
`sparta.Discovery()` **ONLY** works in the context of a Sparta-compatible lambda function, as it depends on Go's reflected function name to perform the CloudFormation lookup.  The initializer creates an inverted index from Go function name to stable CloudFormation resource name.  A successful lookup of a Go function is then used to perform a `cloud formation.DescribeStackResource` call. Assuming the call succeeds, the returned configuration data is the contents of the resources _Metdata_ field.

The contents of the _Metadata_ field are populated during marshaling of the CloudFormation template.  Once all resources have been marshaled to the template, each resource is visited to see if there are any _DependsOn_ relations.  For each _DependsOn_ relation (subsequent commit), the referenced resource's outputs and `Fn:GetAtt` values are copied into the target lambda's _Metadata_.