Move EventSourceMapping IAM annotation to materialize step
EventSourceMapping entries were previously string literals only. Moving
them to dynamic references meant that the IAM policy document couldn't be
fully constructed without all of the resources (and their gocf.StringExpr
references) being defined.

Add support for that via `annotateEventSourceMappings`, which inspects
the aggregated template and EventSourceMapping refs to determine
which policies need to be included in the IAM::Role